ConstantMedium to separate module
