+++ src/hittable/mod.rs
@@ -13,6 +13,7 @@ use rand::RngCore;
 
 pub use aabox::*;
 pub use aarect::*;
+pub use constant_medium::*;
 pub use list::*;
 pub use sphere::*;
 
@@ -27,6 +28,7 @@ mod sphere;
 mod aarect;
 mod list;
 mod aabox;
+mod constant_medium;
 
 #[derive(Copy, Clone)]
 pub struct Hit<'a> {
@@ -234,58 +236,6 @@ impl Hittable for RotateY {
 
     fn bounding_box(&self, t_min: f32, t_max: f32) -> Option<AABB> {
         self.aabb
-    }
-}
-
-
-#[derive(Debug)]
-pub struct ConstantMedium {
-    boundary: Box<dyn Hittable>,
-    density: f64,
-    phase_function: Box<dyn Material>,
-}
-
-impl ConstantMedium {
-    pub fn new(boundary: Box<dyn Hittable>,
-               density: f64,
-               texture: Box<dyn Texture>,
-    ) -> ConstantMedium {
-        ConstantMedium {
-            boundary,
-            density,
-            phase_function: Box::new(Isotropic::new(texture)),
-        }
-    }
-}
-
-impl Hittable for ConstantMedium {
-    fn hit(&self, ray: &Ray, dist_min: f64, dist_max: f64) -> Option<Hit> {
-        self.boundary.hit(ray, 0.0, MAX).and_then(|enter_hit| {
-            self.boundary.hit(ray, enter_hit.dist + 0.001, MAX).and_then(|exit_hit| {
-                let enter_dist = f64::max(dist_min, enter_hit.dist);
-                let exit_dist = f64::min(exit_hit.dist, dist_max);
-                if enter_dist < exit_dist {
-                    // TODO: describe why such distribution?
-                    //  isotropic scattering follows Poisson point process?
-                    let hit_dist = next_f64(rand_distr::Exp1) / self.density;
-                    let inner_travel_distance = (exit_dist - enter_dist) * ray.direction.length();
-                    if hit_dist < inner_travel_distance {
-                        let dist = enter_dist + hit_dist / ray.direction.length();
-                        Some(Hit::new(
-                            dist,
-                            ray.point_at(dist),
-                            V3::new(1.0, 0.0, 0.0),
-                            self.phase_function.borrow(),
-                            0.0, 0.0,
-                        ))
-                    } else { None }
-                } else { None }
-            })
-        })
-    }
-
-    fn bounding_box(&self, t_min: f32, t_max: f32) -> Option<AABB> {
-        self.boundary.bounding_box(t_min, t_max)
     }
 }
